Abstract SoSwitch into SoDevice
