+++ ocf_devices/slined_onboarding/common/__init__.py
@@ -1,3 +1,4 @@
+from slined_onboarding.common.sodevice import SoDevice
 from slined_onboarding.common.so_pi import SoPiUi
 from slined_onboarding.common.so_gpio import SoGpioContext
 from slined_onboarding.common.so_img import SoImgLabel
